+++ jpro-routing/core/src/main/scala/one/jpro/platform/routing/extensions/linkheader/LinkHeaderFilter.scala
@@ -1,7 +1,7 @@
 package one.jpro.platform.routing.extensions.linkheader
 
-import one.jpro.platform.routing.filter.container.ContainerFilter
-import one.jpro.platform.routing.{Filter, LinkUtil, RouteUtils}
+import one.jpro.platform.routing.filter.container.{ContainerFilter, ReactiveContainer}
+import one.jpro.platform.routing.{Filter, LinkUtil}
 
 import collection.JavaConverters._
 import simplefx.core._
@@ -12,33 +12,28 @@ object LinkHeaderFilter {
   case class Link(name: String, prefix: String)
 
   def genLink(name: String, prefix: String) = new Link(name, prefix)
+
   def create(x: Link*): Filter = create(x.toList)
   def create(x: java.util.List[Link]): Filter = create(x.asScala.toList)
-  def create(x: List[Link]): Filter = {
-    ContainerFilter.create(new LinkHeaderFilter.LinkHeaderContainer(x))
-  }
-  private class LinkHeaderContainer(x: List[Link]) extends RouteUtils.SFXContainerFactory {
-    override def isContainer(x: Node): Boolean = x.isInstanceOf[MyContainer]
-    override def createContainer() = new MyContainer
-    class MyContainer extends VBox with Container {
-  
-      this <++  new HBox {
-        styleClass ::= "linkheader-link-hbox"
-        x.map { link =>        
-          this <++ new Label {
-            styleClass ::= "linkheader-link"
-            text <-- link.name
-            when (request != null && request.getPath().startsWith(link.prefix)) ==> {
-              styleClass <++ "selected"
-            }
-            LinkUtil.setLink(this, link.prefix, link.name)
+  def create(x: List[Link]): Filter = ContainerFilter.fromReactiveContainer(() => new LinkHeaderContainer(x))
+
+  private class LinkHeaderContainer(links: List[Link]) extends VBox with ReactiveContainer {
+    this <++ new HBox {
+      styleClass ::= "linkheader-link-hbox"
+      links.map { link =>
+        this <++ new Label {
+          styleClass ::= "linkheader-link"
+          text <-- link.name
+          when(request != null && request.getPath().startsWith(link.prefix)) ==> {
+            styleClass <++ "selected"
           }
+          LinkUtil.setLink(this, link.prefix, link.name)
         }
       }
-      this <++ new StackPane {
-        javafx.scene.layout.VBox.setVgrow(this, Priority.ALWAYS)
-        children <-- (if(content != null) List(content) else Nil)
-      }
+    }
+    this <++ new StackPane {
+      javafx.scene.layout.VBox.setVgrow(this, Priority.ALWAYS)
+      children <-- (if (content != null) List(content) else Nil)
     }
   }
 }